Merged PR 159: STORY-421 As a user, I would like to be able to login
## Pull Request Description (proposed changes, what will be affected):
When unauthenticated, a landing page will be displayed, showing information on the product and call to action buttons to sign into account or activate their account. Unauthenticated users will be able to navigate to other pages but an error message will be displayed. Landing page is responsive. When authenticated, there is now a logout button which redirects to the landing page.

## Testing Instructions:
1. yarn test

## Screenshots/Videos (if not clear):
![image.png](https://dev.azure.com/KPMG-Canada-MC-MS-Practice/0f4dbab5-6200-4134-85d2-d231a994a404/_apis/git/repositories/c6e3598d-3661-4f9e-a3a1-0c1ca121d3e9/pullRequests/159/attachments/image.png)
![image (2).png](https://dev.azure.com/KPMG-Canada-MC-MS-Practice/0f4dbab5-6200-4134-85d2-d231a994a404/_apis/git/repositories/c6e3598d-3661-4f9e-a3a1-0c1ca121d3e9/pullRequests/159/attachments/image%20%282%29.png)

## Checklist:
- [x] Your code builds clean without any errors/warnings
- [x] You are using approved terminology/coding guidelines
- [x] You have added tests
- [x] You did some QA/Validation of your changes

STORY-421 Created LandingPage

Related work items: #421